Move events subscription to useEvents hook
